+++ ss-manager/ss-manager-web/src/main/java/com/qf/web/UserAction.java
@@ -21,7 +21,7 @@ public class UserAction {
      */
     @ResponseBody
     @RequestMapping(value = "/user/{id}",method = RequestMethod.GET)
-    public TbUser findById(@PathVariable("id") Integer id){
+    public TbUser findById(@PathVariable("id") Long id){
         return userService.getById(id);
     }
